requirements.txt
fastapi
Pillow
uvicorn
gunicorn==19.9.0
uvloop
httptools
tensorflow
opencv-python
numpy